Split the GraphQL server creation into separate files for easier testing
