+++ backend/package.json
@@ -17,6 +17,7 @@
   "license": "ISC",
   "dependencies": {
     "dotenv": "^6.0.0",
+    "graphql-import": "^0.7.1",
     "graphql-yoga": "^1.13.1",
     "lodash": "^4.17.10",
     "pg": "^7.4.2",
